UpdateAddHtlc.amountMsat: Long -> UpdateAddHtlc.amountMsat: MilliSatoshi

+++ eclair-core/src/main/scala/fr/acinq/eclair/transactions/CommitmentSpec.scala
@@ -31,7 +31,7 @@ case object OUT extends Direction { def opposite = IN }
 case class DirectedHtlc(direction: Direction, add: UpdateAddHtlc)
 
 final case class CommitmentSpec(htlcs: Set[DirectedHtlc], feeratePerKw: Long, toLocalMsat: Long, toRemoteMsat: Long) {
-  val totalFunds = toLocalMsat + toRemoteMsat + htlcs.toSeq.map(_.add.amountMsat).sum
+  val totalFunds = toLocalMsat + toRemoteMsat + htlcs.toSeq.map(_.add.amountMsat.toLong).sum
 }
 
 object CommitmentSpec {
@@ -43,16 +43,16 @@ object CommitmentSpec {
   def addHtlc(spec: CommitmentSpec, direction: Direction, update: UpdateAddHtlc): CommitmentSpec = {
     val htlc = DirectedHtlc(direction, update)
     direction match {
-      case OUT => spec.copy(toLocalMsat = spec.toLocalMsat - htlc.add.amountMsat, htlcs = spec.htlcs + htlc)
-      case IN => spec.copy(toRemoteMsat = spec.toRemoteMsat - htlc.add.amountMsat, htlcs = spec.htlcs + htlc)
+      case OUT => spec.copy(toLocalMsat = spec.toLocalMsat - htlc.add.amountMsat.toLong, htlcs = spec.htlcs + htlc)
+      case IN => spec.copy(toRemoteMsat = spec.toRemoteMsat - htlc.add.amountMsat.toLong, htlcs = spec.htlcs + htlc)
     }
   }
 
   // OUT means we are sending an UpdateFulfillHtlc message which means that we are fulfilling an HTLC that they sent
   def fulfillHtlc(spec: CommitmentSpec, direction: Direction, htlcId: Long): CommitmentSpec = {
     spec.htlcs.find(htlc => htlc.direction != direction && htlc.add.id == htlcId) match {
-      case Some(htlc) if direction == OUT => spec.copy(toLocalMsat = spec.toLocalMsat + htlc.add.amountMsat, htlcs = spec.htlcs - htlc)
-      case Some(htlc) if direction == IN => spec.copy(toRemoteMsat = spec.toRemoteMsat + htlc.add.amountMsat, htlcs = spec.htlcs - htlc)
+      case Some(htlc) if direction == OUT => spec.copy(toLocalMsat = spec.toLocalMsat + htlc.add.amountMsat.toLong, htlcs = spec.htlcs - htlc)
+      case Some(htlc) if direction == IN => spec.copy(toRemoteMsat = spec.toRemoteMsat + htlc.add.amountMsat.toLong, htlcs = spec.htlcs - htlc)
       case None => throw new RuntimeException(s"cannot find htlc id=${htlcId}")
     }
   }
@@ -60,8 +60,8 @@ object CommitmentSpec {
   // OUT means we are sending an UpdateFailHtlc message which means that we are failing an HTLC that they sent
   def failHtlc(spec: CommitmentSpec, direction: Direction, htlcId: Long): CommitmentSpec = {
     spec.htlcs.find(htlc => htlc.direction != direction && htlc.add.id == htlcId) match {
-      case Some(htlc) if direction == OUT => spec.copy(toRemoteMsat = spec.toRemoteMsat + htlc.add.amountMsat, htlcs = spec.htlcs - htlc)
-      case Some(htlc) if direction == IN => spec.copy(toLocalMsat = spec.toLocalMsat + htlc.add.amountMsat, htlcs = spec.htlcs - htlc)
+      case Some(htlc) if direction == OUT => spec.copy(toRemoteMsat = spec.toRemoteMsat + htlc.add.amountMsat.toLong, htlcs = spec.htlcs - htlc)
+      case Some(htlc) if direction == IN => spec.copy(toLocalMsat = spec.toLocalMsat + htlc.add.amountMsat.toLong, htlcs = spec.htlcs - htlc)
       case None => throw new RuntimeException(s"cannot find htlc id=${htlcId}")
     }
   }
